Fix a compile error due to min overloading

+++ xSchedule/xSMSDaemon/Curl.h
@@ -19,6 +19,7 @@
 #include <wx/url.h>
 
 #include <string>
+#include <algorithm>
 
 #include <log4cpp/Category.hh>
 
@@ -352,7 +353,7 @@ public:
 
                 res = response_string;
                 logger_curl.debug("RESPONSE START ----------");
-                logger_curl.debug(res.substr(0, std::min((size_t)4096, res.size())).c_str());
+                logger_curl.debug(res.substr(0, (std::min)((int)4096, (int)res.size())).c_str());
                 logger_curl.debug("RESPONSE END ----------");
             }
